+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/TimelineViewModel.kt
@@ -146,6 +146,7 @@ interface TimelineViewModelFactory {
  */
 interface TimelineViewModel {
     val elements: StateFlow<List<BaseTimelineElementHolderViewModel>>
+    var firstVisibleElementIndex: StateFlow<Int>
 
     /**
      * Use this to set the state of the current UI.
@@ -235,7 +236,6 @@ class TimelineViewModelImpl(
     private val onOpenMention: OpenMentionCallback,
     private val onOpenMetadata: (eventId: EventId) -> Unit,
 ) : MatrixClientViewModelContext by viewModelContext, TimelineViewModel {
-
     init {
         log.debug { "::: init timelineViewModel: $viewModelContext" }
     }
@@ -338,6 +338,7 @@ class TimelineViewModelImpl(
     override val error: MutableStateFlow<String?> = MutableStateFlow(null)
 
     override val draggedFile: MutableStateFlow<FileDescriptor?> = MutableStateFlow(null)
+    override var firstVisibleElementIndex: StateFlow<Int> = MutableStateFlow(0)
 
     private val directRoom = get<DirectRoom>()
     private val messengerSettings = get<MatrixMessengerSettingsHolder>()
@@ -1062,6 +1063,7 @@ class TimelineViewModelImpl(
 }
 
 class PreviewTimelineViewModel : TimelineViewModel {
+    override var firstVisibleElementIndex: StateFlow<Int> = MutableStateFlow(0)
     override val elements: MutableStateFlow<List<BaseTimelineElementHolderViewModel>> =
         MutableStateFlow(
             listOf(